+++ frontend/app/src/main/java/com/hjy/hackathon/NewChatActivity.kt
@@ -11,9 +11,11 @@ import com.google.firebase.database.ktx.database
 import com.google.firebase.ktx.Firebase
 import com.google.gson.Gson
 import com.hjy.hackathon.adapter.ChatAdapter
-import com.hjy.hackathon.utils.FBAuth
 import com.hjy.hackathon.vo.ChatVO
 import com.hjy.hackathon.vo.MemberVO
+import java.text.SimpleDateFormat
+import java.util.Calendar
+import java.util.Locale
 
 //채팅방 여기서 작업할게요~~
 class NewChatActivity : AppCompatActivity() {
@@ -44,9 +46,8 @@ class NewChatActivity : AppCompatActivity() {
         rvChat.adapter=adapter
 
         btnChatSend.setOnClickListener {
-            val fb = FBAuth;
-            Log.d("time", fb.getTime());
-            myRef.push().setValue(ChatVO(etChatMsg.text.toString(),"수진", fb.myTime(fb.getTime())))
+
+            myRef.push().setValue(ChatVO(etChatMsg.text.toString(),"수진", myTime(getTime())))
 
 
             etChatMsg.text.clear()
@@ -54,4 +55,47 @@ class NewChatActivity : AppCompatActivity() {
 
        myRef.addChildEventListener(ChatChildEvent(data,adapter, rvChat))
     }
+
+    fun getTime(): String {
+        // Calendar 객체는 getInstance() 메소드로 객체를 생성한다
+        val currentTime = Calendar.getInstance().time
+        // 시간을 나타낼 형식, 어느위치의 시간을 가져올건지 설정
+        // "yyyy.MM.dd HH:mm:sss"
+
+        val time = SimpleDateFormat("yyyy.MM.dd HH:mm", Locale.KOREAN).format(currentTime)
+
+        return time
+    }
+
+    /**시간을 오후 9:13같은 형식으로 바꿔준다*/
+    fun myTime(time: String): String {
+        var timeResult = ""
+        if (time == "")
+            return ""
+        if (time.substring(11, 13).toInt() > 12)
+            timeResult += "오후 " + (time.substring(11, 13).toInt() - 12) + time.substring(13, 16)
+        else if (time.substring(11, 13).toInt() == 12)
+            timeResult += "오후 " + time.substring(11, 16)
+        else if (time.substring(11, 13).toInt() > 9)
+            timeResult += "오전 " + time.substring(11, 16)
+        else if (time.substring(11, 13).toInt() == 0)
+            timeResult += "오전 12" + time.substring(13, 16)
+        else
+            timeResult += "오전 " + time.substring(12, 16)
+        return timeResult
+    }
+
+//    // 아이디 비교하기
+//
+//    /**사용자uid와 targetUid 비교
+//     * @return Boolean*/
+//    fun checkUid(targetUid: String): Boolean {
+//        val userUid = FBAuth.getUid()
+//        Log.d("check", " 사용자uid: $userUid, targetUid: $targetUid")
+//        return targetUid == userUid
+//    }
+
+
+
+
 }
