--- frontend/app/src/main/res/xml/network_security_config.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<network-security-config>
- <domain-config cleartextTrafficPermitted="true">
-     <domain includeSubdomains="true">your_site_domain.com</domain>
- </domain-config>
-</network-security-config>
